Introduce PongContainer
